requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
pdf2image
Pillow
openai
pyjwt
python-dotenv
